[IMP] f&r: remove find & replace plugin
Now that we can batch commands, the plugin find & replace
is now useless.

Task: 3870119
