+++ src/types/commands.ts
@@ -11,7 +11,6 @@ import {
 import {
   Border,
   BorderData,
-  CellPosition,
   Dimension,
   HeaderIndex,
   Pixel,
@@ -23,7 +22,6 @@ import {
 import { ChartDefinition } from "./chart/chart";
 import { ClipboardPasteOptions } from "./clipboard";
 import { FigureSize } from "./figure";
-import { SearchOptions } from "./find_and_replace";
 import { Image } from "./image";
 import { PivotCoreDefinition, SPTableData } from "./pivot";
 import { RangeData } from "./range";
@@ -825,14 +823,6 @@ export interface SelectFigureCommand {
   id: UID | null;
 }
 
-export interface ReplaceSearchCommand {
-  type: "REPLACE_SEARCH";
-  searchString: string;
-  replaceWith: string;
-  searchOptions: SearchOptions;
-  matches: CellPosition[];
-}
-
 export interface SortCommand {
   type: "SORT_CELLS";
   sheetId: UID;
@@ -1052,7 +1042,6 @@ export type LocalCommand =
   | ShowFormulaCommand
   | AutofillAutoCommand
   | SelectFigureCommand
-  | ReplaceSearchCommand
   | SortCommand
   | SetDecimalCommand
   | ResizeViewportCommand
